Add Litigation Cases exception for USA
I also removed the filter on links on the profile sidebar when there's
no data for a type of data, as I think this is more consistent and there
might be different reasons to not have data.
Maybe we could add a field to the countries to "Why no Litigation DAta"?
:D but I won't do it here now.